enable nullable fields for userinfo

+++ apps/services/user-profile/src/app/user-profile/userProfile.controller.ts
@@ -7,7 +7,6 @@ import {
   Param,
   Post,
   ConflictException,
-  BadRequestException,
 } from '@nestjs/common'
 import {
   ApiCreatedResponse,
@@ -34,7 +33,7 @@ export class UserProfileController {
   constructor(
     private userProfileService: UserProfileService,
     private verificationService: VerificationService,
-  ) {}
+  ) { }
 
   @Get('userProfile/:nationalId')
   @ApiParam({
@@ -65,10 +64,12 @@ export class UserProfileController {
       )
     }
 
-    await this.verificationService.createEmailVerification(
-      userProfileDto.nationalId,
-      userProfileDto.email,
-    )
+    if (userProfileDto.email) {
+      await this.verificationService.createEmailVerification(
+        userProfileDto.nationalId,
+        userProfileDto.email,
+      )
+    }
 
     if (userProfileDto.mobilePhoneNumber) {
       const phoneVerified = await this.verificationService.isPhoneNumberVerified(
@@ -78,13 +79,14 @@ export class UserProfileController {
         ...userProfileDto,
         mobilePhoneNumberVerified: phoneVerified,
       }
+      if (phoneVerified) {
+        await this.verificationService.removeSmsVerification(
+          userProfileDto.nationalId,
+        )
+      }
     }
-    const profile = await this.userProfileService.create(userProfileDto)
-    await this.verificationService.removeSmsVerification(
-      userProfileDto.nationalId,
-    )
 
-    return profile
+    return await this.userProfileService.create(userProfileDto)
   }
 
   @Put('userProfile/:nationalId')
@@ -119,7 +121,7 @@ export class UserProfileController {
       }
     }
 
-    if (userProfileToUpdate.email !== profile.email) {
+    if (userProfileToUpdate.email && userProfileToUpdate.email !== profile.email) {
       await this.verificationService.createEmailVerification(
         nationalId,
         userProfileToUpdate.email,
